Issue: #129 #122 #124 Moved the notifications related stuff to the navbar instead of homepage so can be accessed on any page Time: 2 hours
