+++ src/app/components/app.scss
@@ -484,3 +484,7 @@ hr {
 .spacingTopRight{
   margin-left: 92%;
 }
+
+.gold{
+  color: gold !important;
+}
